custom-elements/registering-custom-elements/definition-construction-algorithm-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type should be lowercased</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Custom element type should be lowercased">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var GeneratedConstructor = doc.register('X-A-1');
    var customElement = new GeneratedConstructor();

    assert_equals(customElement.localName, 'x-a-1', 'Custom element type should be lowercased');

}, 'Custom element type should be lowercased. Test constructor');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function (element) {
        var obj = doc.createElement(element);        
        var name = 'x-' + element;
        var proto = Object.create(obj.constructor.prototype);
        var GeneratedConstructor = doc.register(name, {
            prototype: proto, 
            extends: element.toUpperCase()
        });
        var customElement = new GeneratedConstructor();
            
        assert_equals(customElement.localName, element, 'Local name should be lowercased');
    });
}, 'Custom element type should be lowercased. Test constructor of extended HTML element');

</script>
</body>
</html>
